refactor(agent-ui): inline ChatPane's single-caller wrappers and the router.push alias
useChatPaneComposerActions carried one useCallback for one caller; the
diff-drawer/subagent-chip/pi-session-id render hops each hid a one-line
conditional; navigateToSessionHref was a router.push alias. All now live
at their call sites.
